Finalize assignment1

+++ array.h
@@ -24,7 +24,7 @@ class Array : public Object {
          * @param index The index at which the item will be retreived from
          * @return The object residing at the index
          */
-        Object* get(int index) { }
+        Object* get(size_t index) { }
 
         /**
          * Adds an item to the start of an array
@@ -44,37 +44,37 @@ class Array : public Object {
          * @param index The index at which the item will be inserted
          * @return The object that was replaced
          */
-        Object* set(Object* item, int index) { }
+        Object* set(Object* item, size_t index) { }
 
         /**
          * Inserts the element at the given index
          * @param item The item to be inserted
          * @param index The index at which the item will be inserted
          */
-        void insert(Object* item, int index) { }
+        void insert(Object* item, size_t index) { }
 
         /**
          * Removes the element at the given index
          * @param index The index at which the item will be removed
          * @return The object that was removed
          */
-        Object* remove(int index) { }
+        Object* remove(size_t index) { }
 
         /**
          * Returns the length of the array
          * @return length of the array
          */
-        int length() { }
+        size_t length() { }
 
         /**
          * Returns if the array contains the item
          * @return 1 if contains item, 0 otherwise
          */
-        int contains(Object* item) { }
+        size_t contains(Object* item) { }
 
         /**
          * Returns the index of the item
          * @return index of the item. -1 if it does not contain the item
          */
-        int indexOf(Object* item) { }
+        size_t indexOf(Object* item) { }
 };
